+++ lib/app/theme/app_theme.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class MyThemes {
+  static final darkTheme = ThemeData(
+    scaffoldBackgroundColor: Colors.black,
+    primaryColor: Colors.black,
+    colorScheme: const ColorScheme.dark(),
+  );
+
+  static final lightTheme = ThemeData(
+    scaffoldBackgroundColor: Colors.white,
+    primaryColor: Colors.white,
+    colorScheme: const ColorScheme.light(),
+  );
+}
