+++ lib/app/utils/snackbar/snackbar_ext.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+extension SnackBarExt on BuildContext {
+  void showSnackBar(String message, {bool isError = false}) =>
+      ScaffoldMessenger.of(this).showSnackBar(
+        SnackBar(
+          content: Text(message),
+          duration: const Duration(seconds: 2),
+          backgroundColor:
+              isError ? Colors.red.shade400 : Colors.green.shade400,
+        ),
+      );
+
+  void hideSnackBar(
+          {SnackBarClosedReason reason = SnackBarClosedReason.hide}) =>
+      ScaffoldMessenger.of(this).hideCurrentSnackBar(reason: reason);
+
+  void clearAllSnackBars() => ScaffoldMessenger.of(this).clearSnackBars();
+
+  void showSnackBarWithAction(
+    String message, {
+    required String actionLabel,
+    required VoidCallback onAction,
+    bool isError = false,
+  }) =>
+      ScaffoldMessenger.of(this).showSnackBar(
+        SnackBar(
+          content: Text(message),
+          duration: const Duration(seconds: 2),
+          backgroundColor:
+              isError ? Colors.red.shade400 : Colors.green.shade400,
+          action: SnackBarAction(
+            label: actionLabel,
+            onPressed: onAction,
+          ),
+        ),
+      );
+}
