--- lib/providers/theme_provider.dart
@@ -1,34 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
-
-class ThemeNotifier extends ChangeNotifier {
-  ThemeMode themeMode = ThemeMode.system;
-
-  bool get isDarkMode {
-    if (themeMode == ThemeMode.system) {
-      final brightness = SchedulerBinding.instance.window.platformBrightness;
-      return brightness == Brightness.dark;
-    } else {
-      return themeMode == ThemeMode.dark;
-    }
-  }
-
-  void toggleTheme(bool isOn) {
-    themeMode = isOn ? ThemeMode.dark : ThemeMode.light;
-    notifyListeners();
-  }
-}
-
-class MyThemes {
-  static final darkTheme = ThemeData(
-    scaffoldBackgroundColor: Colors.black,
-    primaryColor: Colors.black,
-    colorScheme: const ColorScheme.dark(),
-  );
-
-  static final lightTheme = ThemeData(
-    scaffoldBackgroundColor: Colors.white,
-    primaryColor: Colors.white,
-    colorScheme: const ColorScheme.light(),
-  );
-}
