--- lib/screens/splashscreen.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:animated_splash_screen/animated_splash_screen.dart';
-import 'package:page_transition/page_transition.dart';
-import 'package:watch/screens/signup_page.dart';
-
-class SplashScreen extends StatelessWidget {
-  const SplashScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedSplashScreen(
-      backgroundColor: Colors.black45,
-      splash: 'assets/logo.png',
-      splashIconSize: 200,
-      nextScreen: SignUpPage(),
-      splashTransition: SplashTransition.sizeTransition,
-      pageTransitionType: PageTransitionType.bottomToTop,
-    );
-  }
-}
