+++ lib/src/features/authentication/controller/auth_controller.dart
@@ -1,0 +1,1 @@
+//TODO: Write the auth controllers here
