+++ lib/src/features/home/presentation/home_view/components/explore.dart
