+++ lib/src/features/home/presentation/home_view/components/fulldetails.dart
