+++ lib/src/features/splash/presentation/splash_screen.dart
@@ -1,0 +1,20 @@
+import 'package:animated_splash_screen/animated_splash_screen.dart';
+import 'package:flutter/material.dart';
+import 'package:page_transition/page_transition.dart';
+import 'package:watch/src/features/authentication/presentation/signup/signup_screen.dart';
+
+class SplashScreen extends StatelessWidget {
+  const SplashScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimatedSplashScreen(
+      backgroundColor: Colors.black45,
+      splash: 'assets/logo.png',
+      splashIconSize: 200,
+      nextScreen: SignUpPage(),
+      splashTransition: SplashTransition.sizeTransition,
+      pageTransitionType: PageTransitionType.bottomToTop,
+    );
+  }
+}
