+++ lib/src/shared/custom_paints.dart
