--- lib/util/theme.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-
-final lightTheme = ThemeData(
-    brightness: Brightness.light,
-    textTheme: GoogleFonts.interTextTheme().apply());
-final darkTheme = ThemeData(
-    brightness: Brightness.dark,
-    scaffoldBackgroundColor: Colors.black,
-    textTheme: GoogleFonts.interTextTheme().apply());
-
-class ThemeNotifier with ChangeNotifier {
-  ThemeData _themeData;
-
-  ThemeNotifier(this._themeData);
-  getTheme() => _themeData;
-
-  setTheme(ThemeData themeData) async {
-    _themeData = themeData;
-    notifyListeners();
-  }
-}
